Refactored CloudinaryConfig to remove circular dependency
